launch-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
PyJWT
requests
python-jose[cryptography]
httpx